--- VocaDbModel/Mapping/Albums/AlbumWebLinkMap.cs
@@ -1,6 +1,0 @@
-﻿using VocaDb.Model.Domain.Albums;
-
-namespace VocaDb.Model.Mapping.Albums
-{
-	public class AlbumWebLinkMap : WebLinkMap<AlbumWebLink, Album> { }
-}
