--- VocaDbModel/Mapping/Artists/ArtistWebLinkMap.cs
@@ -1,6 +1,0 @@
-﻿using VocaDb.Model.Domain.Artists;
-
-namespace VocaDb.Model.Mapping.Artists
-{
-	public class ArtistWebLinkMap : WebLinkMap<ArtistWebLink, Artist> { }
-}
