--- VocaDbModel/Mapping/Users/UserWebLinkMap.cs
@@ -1,9 +1,0 @@
-﻿using VocaDb.Model.Domain.Users;
-
-namespace VocaDb.Model.Mapping.Users
-{
-	public class UserWebLinkMap : WebLinkMap<UserWebLink, User>
-	{
-		public UserWebLinkMap() : base(false) { }
-	}
-}
